# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  # Do not change the title, if the title changes, the import paths will be broken
  version: 0.2.0
  description: AdsGram Telegram Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mini/games/spin/rooms:
    get:
      operationId: getMiniSpinRooms
      tags: [mini-games]
      summary: List open spin rooms
      responses:
        "200":
          description: Spin rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiniSpinRoomList"
    post:
      operationId: createMiniSpinRoom
      tags: [mini-games]
      summary: Create a new PvP spin room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MiniSpinRoomInput"
      responses:
        "200":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiniSpinRoom"
  /mini/games/spin/rooms/{id}:
    get:
      operationId: getMiniSpinRoom
      tags: [mini-games]
      summary: Get spin room details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spin room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiniSpinRoom"
  /mini/games/spin/rooms/{id}/join:
    post:
      operationId: joinMiniSpinRoom
      tags: [mini-games]
      summary: Join a PvP spin room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MiniSpinJoinInput"
      responses:
        "200":
          description: Joined room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiniSpinRoom"
...
components:
  schemas:
...
    MiniSpinPlayer:
      type: object
      properties:
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        stake:
          type: integer
        chance:
          type: number
      required: [telegramId, stake, chance]
    MiniSpinRoom:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        totalPool:
          type: integer
        winnerId:
          type: ["string", "null"]
        winnerUsername:
          type: ["string", "null"]
        players:
          type: array
          items:
            $ref: "#/components/schemas/MiniSpinPlayer"
        startAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, status, totalPool, players, createdAt]
    MiniSpinRoomList:
      type: object
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/MiniSpinRoom"
      required: [rooms]
    MiniSpinRoomInput:
      type: object
      properties:
        telegramId:
          type: string
        stake:
          type: integer
      required: [telegramId, stake]
    MiniSpinJoinInput:
      type: object
      properties:
        telegramId:
          type: string
        stake:
          type: integer
      required: [telegramId, stake]
...